+++ src/paintServer.ts
@@ -4,23 +4,27 @@ import tinycolor from "tinycolor2";
 import { traverse } from "./svgConstructor";
 
 export type PaintServer = {
-    kind: "linearGradient",
+    kind: "linearGradient" | "radialGradient",
     stops: StopReference[]
 }
 
 export function fetchPaintServer(uuid: string): PaintServer | null {
     const pe = svgVirtualMap[uuid];
-    switch (pe.tag) {
-        case "linearGradient":
+    function gradient(pe: ParsedElement & {children: ParsedElement[]} & {tag: "linearGradient" | "radialGradient"}) {
         const stops: StopReference[] = [];
         for (let c of pe.children) {
             let tmp = fetchStopReference(c);
             if (tmp) stops.push(tmp);
         }
         return {
-            kind: "linearGradient",
+            kind: pe.tag,
             stops
         };
+    }
+    switch (pe.tag) {
+        case "linearGradient":
+        case "radialGradient":
+        return gradient(pe);
         default:
         return null;
     }
@@ -54,17 +58,25 @@ export function fetchStopReference(pe: ParsedElement): StopReference | null {
 }
 
 export function cssString(paintServer: PaintServer): string {
-    switch (paintServer.kind) {
-        case "linearGradient":
+    function gradient(paintServer: PaintServer & {stops: StopReference[]}) {
         const acc: string[] = [];
         for (let stop of paintServer.stops) {
             const stopColor = stop["stop-color"];
             acc.push([
-                isColor(stopColor) ? tinycolor(stopColor).toString(stopColor.format) : isFuncIRI(stopColor) ? `url(${stopColor.url})` : stopColor,
+                isColor(stopColor) ? tinycolor(stopColor).toString(stopColor.format) : stopColor,
                 `${stop.offset.value}%`
             ].join(" "));
         }
+        return acc;
+    }
+    let acc: string[];
+    switch (paintServer.kind) {
+        case "linearGradient":
+        acc = gradient(paintServer);
         return `linear-gradient(to right, ${acc.join(", ")})`;
+        case "radialGradient":
+        acc = gradient(paintServer);
+        return `radial-gradient(${acc.join(", ")})`;
     }
 }
 
@@ -72,7 +84,7 @@ export function collectPaintServer(pe: ParsedElement): {[id: string]: ParsedElem
     const acc: {[id: string]: ParsedElement} = {};
     traverse(pe, (pe, parentPe, index) => {
         const ident = pe.attrs.id;
-        if (ident && pe.tag === "linearGradient") {
+        if (ident && (pe.tag === "linearGradient" || pe.tag === "radialGradient")) {
             acc[ident] = pe;
         }
     });