+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 - 2.2.0
   - Update only differences when modifying SVG code
-  - Support linearGradient again
+  - Support linearGradient, radialGradient again
 - 2.1.0
   - Fix auto deletion unsupported attributes bug
   - add basic font selection window